+++ App/canary/commands/admin/dm-remover.js
@@ -4,44 +4,44 @@ module.exports = {
     category: 'admin',
     data: new SlashCommandBuilder()
         .setName('dm-remover')
-        .setDescription('Delete a specific direct message sent by the APP by its ID.')
+        .setDescription('Delete a specific direct message sent by the APP by its ID!')
         .setDefaultMemberPermissions(0)
         .addStringOption(option => option
             .setName('message_id')
-            .setDescription('Message ID to delete')
+            .setDescription('Message ID to delete!')
             .setRequired(true),
         ),
     async execute(interaction) {
         if (interaction.guild) {
-            interaction.reply({ content: 'This command can only be used in direct messages.', ephemeral: true });
+            return interaction.reply({ content: 'This command can only be used in direct messages.', ephemeral: true });
         }
-        else {
-            const messageId = interaction.options.getString('message_id');
-            if (!/^\d+$/.test(messageId)) {
-                interaction.reply({ content: 'Please provide a valid message ID.', ephemeral: true });
+
+        const messageId = interaction.options.getString('message_id');
+
+        if (!/^\d+$/.test(messageId)) {
+            return interaction.reply({ content: 'Please provide a valid message ID.', ephemeral: true });
+        }
+
+        try {
+            const channel = await interaction.user.createDM();
+            const message = await channel.messages.fetch(messageId);
+
+            if (message && message.author.id === interaction.client.user.id) {
+                await message.delete();
+                return interaction.reply({ content: 'The message was successfully deleted.', ephemeral: true });
             }
             else {
-                try {
-                    const channel = await interaction.user.createDM();
-                    const message = await channel.messages.fetch(messageId);
-                    if (message && message.author.id === interaction.client.user.id) {
-                        await message.delete();
-                        interaction.reply({ content: 'The message was successfully deleted.', ephemeral: true });
-                    }
-                    else {
-                        interaction.reply({ content: 'Message not found or cannot be deleted.', ephemeral: true });
-                    }
-                }
-                catch (error) {
-                    // DiscordAPIError[10008]: Unknown Message
-                    if (error.code === 10008) {
-                        await interaction.reply({ content: 'Message not found or cannot be deleted.', ephemeral: true });
-                    }
-                    else {
-                        console.error('Error deleting message:', error);
-                        await interaction.reply({ content: 'There was an error trying to delete the message.', ephemeral: true });
-                    }
-                }
+                return interaction.reply({ content: 'Message not found or cannot be deleted.', ephemeral: true });
+            }
+        }
+        catch (error) {
+            // DiscordAPIError[10008]: Unknown Message
+            if (error.code === 10008) {
+                return interaction.reply({ content: 'Message not found or cannot be deleted.', ephemeral: true });
+            }
+            else {
+                console.error('Error deleting message:', error);
+                return interaction.reply({ content: 'There was an error trying to delete the message.', ephemeral: true });
             }
         }
     },