+++ App/canary/commands/admin/reload.js
@@ -4,26 +4,27 @@ module.exports = {
     category: 'admin',
     data: new SlashCommandBuilder()
         .setName('reload')
-        .setDescription('Reloads a command.')
+        .setDescription('Reloads a command!')
         .setDefaultMemberPermissions(0)
         .setDMPermission(false)
-        .addStringOption(option =>
-            option.setName('command')
-                .setDescription('The command to reload.')
-                .setRequired(true)),
+        .addStringOption(option => option
+            .setName('command')
+            .setDescription('The command to reload!')
+            .setRequired(true),
+        ),
     async execute(interaction) {
         const commandName = interaction.options.getString('command', true).toLowerCase();
         const command = interaction.client.commands.get(commandName);
 
         if (!command) {
-            return interaction.reply({ content: `There is no command with name \`/${commandName}\`!`, ephemeral: true });
+            return interaction.reply({ content: `There is no command with name \`/${commandName}\``, ephemeral: true });
         }
 
-        delete require.cache[require.resolve(`../${command.category}/${command.data.name}.js`)];
+        delete require.cache[require.resolve(`./../${command.category}/${command.data.name}.js`)];
 
         try {
             interaction.client.commands.delete(command.data.name);
-            const newCommand = require(`../${command.category}/${command.data.name}.js`);
+            const newCommand = require(`./../${command.category}/${command.data.name}.js`);
             interaction.client.commands.set(newCommand.data.name, newCommand);
             await interaction.reply({ content: `Command \`/${newCommand.data.name}\` was reloaded!`, ephemeral: true });
         }