+++ App/canary/commands/context-menu/kick.js
@@ -1,12 +1,15 @@
-const { ContextMenuCommandBuilder, ApplicationCommandType, PermissionFlagsBits } = require('discord.js');
+const {
+    ContextMenuCommandBuilder,
+    ApplicationCommandType,
+} = require('discord.js');
 
 module.exports = {
     category: 'context-menu',
     data: new ContextMenuCommandBuilder()
         .setName('Kick')
         .setType(ApplicationCommandType.User)
-        .setDefaultMemberPermissions(PermissionFlagsBits.KickMembers),
+        .setDMPermission(false),
     async execute(interaction) {
-        return interaction.reply({ content: `You wanted to kick: ${interaction.targetUser} :scales:`, ephemeral: true });
+        await interaction.reply({ content: `You wanted to kick: ${interaction.targetUser} :scales:`, ephemeral: true });
     },
 };