+++ App/src/commands/context-menu/user.js
@@ -1,4 +1,8 @@
-const { ContextMenuCommandBuilder, ApplicationCommandType, EmbedBuilder } = require('discord.js');
+const {
+    ContextMenuCommandBuilder,
+    ApplicationCommandType,
+    EmbedBuilder,
+} = require('discord.js');
 
 module.exports = {
     category: 'context-menu',
@@ -16,6 +20,7 @@ module.exports = {
                 { name: 'ID:', value: `${user.id}`, inline: true },
                 { name: 'Joined at:', value: `${member.joinedAt}`, inline: false },
             );
+
         return interaction.reply({ embeds: [embed], ephemeral: true });
     },
 };